fix: fixed install location & silent install
Reverted to having a fixed install location in %LOCALAPPDATA%. Having a user being able to change this will break the bat file and also causes issues with WaspLauncher. This has also proven to be an ineffective change for false positives.

Added a --silent flag for silent installs for 100% unattended bat file force updates.